feat: render command groups
